# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Collaborative Permaculture Design API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /properties/{propertyId}/client-brief:
    get:
      operationId: getClientBrief
      tags: [client-brief]
      summary: Get the client brief / site survey for a property
      parameters:
        - name: propertyId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Client brief
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientBrief"
        "404":
          description: Not found

    put:
      operationId: upsertClientBrief
      tags: [client-brief]
      summary: Create or update the client brief for a property
      parameters:
        - name: propertyId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientBriefInput"
      responses:
        "200":
          description: Saved client brief
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientBrief"
  /properties/{propertyId}/pathways:
    get:
      operationId: listPathways
      tags: [pathways]
      summary: List all access pathways for a property
      parameters:
        - name: propertyId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of pathways
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Pathway"

    post:
      operationId: createPathway
      tags: [pathways]
      summary: Save an access pathway
      parameters:
        - name: propertyId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PathwayInput"
      responses:
        "201":
          description: Created pathway
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pathway"
  /properties/{propertyId}/pathways/{pathwayId}:
    delete:
      operationId: deletePathway
      tags: [pathways]
      summary: Delete an access pathway
      parameters:
        - name: propertyId
          in: path
          required: true
          schema:
            type: string
        - name: pathwayId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
  /properties/{propertyId}/comments:
    get:
      operationId: listComments
      tags: [comments]
      summary: List all feedback pins for a property
      parameters:
        - name: propertyId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"

    post:
      operationId: createComment
      tags: [comments]
      summary: Add a feedback pin to a property
      parameters:
        - name: propertyId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "201":
          description: Created comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
  /properties/{propertyId}/comments/{commentId}:
    delete:
      operationId: deleteComment
      tags: [comments]
      summary: Delete a feedback pin
      parameters:
        - name: propertyId
          in: path
          required: true
          schema:
            type: string
        - name: commentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
  /auth/user:
    get:
      tags: [Auth]
      operationId: getCurrentAuthUser
      summary: Get the currently authenticated user
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Auth status resolved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserEnvelope'
  /login:
    get:
      tags: [Auth]
      operationId: beginBrowserLogin
      summary: Start the browser OIDC login flow
      parameters:
        - name: returnTo
          in: query
          required: false
          schema:
            type: string
      responses:
        '302':
          description: Redirect to the OIDC authorization endpoint.
  /callback:
    get:
      tags: [Auth]
      operationId: handleBrowserLoginCallback
      summary: Complete the browser OIDC login flow
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
      responses:
        '302':
          description: Redirect to the returnTo path on success.
  /logout:
    get:
      tags: [Auth]
      operationId: logoutBrowserSession
      summary: Clear the session and begin OIDC logout
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '302':
          description: Redirect to the OIDC provider logout URL.
  /mobile-auth/token-exchange:
    post:
      tags: [Auth]
      operationId: exchangeMobileAuthorizationCode
      summary: Exchange a mobile OIDC code for a session token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MobileTokenExchangeRequest'
      responses:
        '200':
          description: Session created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileTokenExchangeSuccess'
        '400':
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Token did not contain usable claims.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Token exchange failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /mobile-auth/logout:
    post:
      tags: [Auth]
      operationId: logoutMobileSession
      summary: Delete a mobile session token
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
      responses:
        '200':
          description: Logout processed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutSuccess'
...
components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Opaque session token — Bearer <sid>.
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string
  schemas:
...
    ClientBrief:
      type: object
      required: [id, propertyId, machineryWidthM, utilitiesOverheadPower, utilitiesBuriedPipes, utilitiesLegalEasements, utilitiesActiveWell, challengeSevereErosion, challengeWinterFlooding, challengeHighWind, challengeWildlifePressure, createdAt, updatedAt]
      properties:
        id:
          type: string
        propertyId:
          type: string
        annualRainfallMm:
          type: number
          nullable: true
        estimatedSoilType:
          type: string
          nullable: true
        climateZone:
          type: string
          nullable: true
        # Extended site baseline fields
        elevationM:
          type: number
          nullable: true
        meanAnnualTempC:
          type: number
          nullable: true
        summerMaxTempC:
          type: number
          nullable: true
        winterMinTempC:
          type: number
          nullable: true
        frostDaysPerYear:
          type: number
          nullable: true
        annualHumidityPct:
          type: number
          nullable: true
        prevailingWindDir:
          type: string
          nullable: true
        meanWindSpeedMs:
          type: number
          nullable: true
        solarIrradianceKwhM2:
          type: number
          nullable: true
        soilClay:
          type: number
          nullable: true
        soilSand:
          type: number
          nullable: true
        soilSilt:
          type: number
          nullable: true
        soilPH:
          type: number
          nullable: true
        soilOrganicCarbonGkg:
          type: number
          nullable: true
        soilTextureClass:
          type: string
          nullable: true
        machineryWidthM:
          type: number
        utilitiesOverheadPower:
          type: boolean
        utilitiesBuriedPipes:
          type: boolean
        utilitiesLegalEasements:
          type: boolean
        utilitiesActiveWell:
          type: boolean
        challengeSevereErosion:
          type: boolean
        challengeWinterFlooding:
          type: boolean
        challengeHighWind:
          type: boolean
        challengeWildlifePressure:
          type: boolean
        primaryGoal:
          type: string
          nullable: true
        householdSize:
          type: integer
          nullable: true
        maintenanceCapacity:
          type: string
          nullable: true
        moodBoardImages:
          type: array
          nullable: true
          items:
            type: string
        conceptRenders:
          type: array
          nullable: true
          items:
            type: string
        aiAnalysisReport:
          type: string
          nullable: true
        aiAnalysisGeneratedAt:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
    ClientBriefInput:
      type: object
      required: [machineryWidthM, utilitiesOverheadPower, utilitiesBuriedPipes, utilitiesLegalEasements, utilitiesActiveWell, challengeSevereErosion, challengeWinterFlooding, challengeHighWind, challengeWildlifePressure]
      properties:
        annualRainfallMm:
          type: number
          nullable: true
        estimatedSoilType:
          type: string
          nullable: true
        climateZone:
          type: string
          nullable: true
        elevationM:
          type: number
          nullable: true
        meanAnnualTempC:
          type: number
          nullable: true
        summerMaxTempC:
          type: number
          nullable: true
        winterMinTempC:
          type: number
          nullable: true
        frostDaysPerYear:
          type: number
          nullable: true
        annualHumidityPct:
          type: number
          nullable: true
        prevailingWindDir:
          type: string
          nullable: true
        meanWindSpeedMs:
          type: number
          nullable: true
        solarIrradianceKwhM2:
          type: number
          nullable: true
        soilClay:
          type: number
          nullable: true
        soilSand:
          type: number
          nullable: true
        soilSilt:
          type: number
          nullable: true
        soilPH:
          type: number
          nullable: true
        soilOrganicCarbonGkg:
          type: number
          nullable: true
        soilTextureClass:
          type: string
          nullable: true
        machineryWidthM:
          type: number
        utilitiesOverheadPower:
          type: boolean
        utilitiesBuriedPipes:
          type: boolean
        utilitiesLegalEasements:
          type: boolean
        utilitiesActiveWell:
          type: boolean
        challengeSevereErosion:
          type: boolean
        challengeWinterFlooding:
          type: boolean
        challengeHighWind:
          type: boolean
        challengeWildlifePressure:
          type: boolean
        primaryGoal:
          type: string
          nullable: true
        householdSize:
          type: integer
          nullable: true
        maintenanceCapacity:
          type: string
          nullable: true
        moodBoardImages:
          type: array
          nullable: true
          items:
            type: string
        conceptRenders:
          type: array
          nullable: true
          items:
            type: string
...
    Pathway:
      type: object
      required: [id, propertyId, label, pathwayType, lineGeojson, createdAt]
      properties:
        id:
          type: string
        propertyId:
          type: string
        label:
          type: string
        pathwayType:
          type: string
        lineGeojson:
          type: string
        createdAt:
          type: string
    PathwayInput:
      type: object
      required: [label, pathwayType, lineGeojson]
      properties:
        label:
          type: string
          minLength: 1
        pathwayType:
          type: string
        lineGeojson:
          type: string
...
    Comment:
      type: object
      required: [id, propertyId, lng, lat, text, authorRole, createdAt]
      properties:
        id:
          type: string
        propertyId:
          type: string
        lng:
          type: number
        lat:
          type: number
        text:
          type: string
        authorRole:
          type: string
        createdAt:
          type: string
    CommentInput:
      type: object
      required: [lng, lat, text, authorRole]
      properties:
        lng:
          type: number
        lat:
          type: number
        text:
          type: string
        authorRole:
          type: string
    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl, isStaff]
      properties:
        id:
          type: string
        email:
          type: [string, 'null']
          format: email
        firstName:
          type: [string, 'null']
        lastName:
          type: [string, 'null']
        profileImageUrl:
          type: [string, 'null']
        isStaff:
          type: boolean
    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: '#/components/schemas/AuthUser'
            - type: 'null'
    MobileTokenExchangeRequest:
      type: object
      required: [code, code_verifier, redirect_uri, state]
      properties:
        code:
          type: string
          minLength: 1
        code_verifier:
          type: string
          minLength: 1
        redirect_uri:
          type: string
          minLength: 1
          format: uri
        state:
          type: string
          minLength: 1
        nonce:
          type: string
          minLength: 1
    MobileTokenExchangeSuccess:
      type: object
      required: [token]
      properties:
        token:
          type: string
    LogoutSuccess:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string